pyproject.toml
[tool.poetry]
name = "machine-learning-playground"
version = "0.1.0"
description = ""
authors = ["skewballfox <joshua.ferguson.273@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "machine_learning_playground"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.23.3"
sklearn = "^0.0"
scipy = "^1.9.1"
matplotlib = "^3.6.0"
pandas = "^1.5.0"
Pillow = "^9.2.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
